Fix to model

+++ 05-postgresql/models/models.go
@@ -1,8 +1,8 @@
 package models
 
 type Stock struct {
-	StockId int64  `json:"stockId"`
-	Name    string `json:"name"`
-	Price   int    `json:"price"`
-	Company string `json:"company"`
+	StockId int64   `json:"stockId"`
+	Name    string  `json:"name"`
+	Price   float64 `json:"price"`
+	Company string  `json:"company"`
 }
